+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 FEATURES:
 * redis: added `enable_sharding` feature, it works when cluster was not sharded and new value of sharded is true
 * storage: support `tags` and `and` operator in bucket lifecycle configuration in `yandex_storage_bucket` resource
+* **New Resource:** `yandex_mdb_mongodb_database`
+* **New Resource:** `yandex_mdb_mongodb_user`
+* **New Data Source:** `yandex_mdb_mongodb_database`
+* **New Data Source:** `yandex_mdb_mongodb_user`
+
 BUG FIXES:
 * fix ydb_table type
 * serverless: fix YDS trigger creation.
+
+WARNING:
+* mongodb: `database` and `user` sections for `yandex_mdb_postgresql_cluster` are now deprecated
 
 ## 0.107.0 (January 31, 2024)
 FEATURES: